admin_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
httpx==0.27.0
redis==5.0.3
pydantic==2.7.1
